Move Google OAuth for device authentication to google_auth module
