+++ flirror/__init__.py
@@ -1,11 +1,5 @@
 from flask import Flask
 
-from .google_auth import (
-    OAuth2CallbackView,
-    Oauth2ClearView,
-    OAuth2RevokeView,
-    OAuth2View,
-)
 from .helpers import make_error_handler
 from .utils import weather_icon
 from .views import CalendarView, IndexView, MapView, WeatherView
@@ -26,12 +20,6 @@ def create_app():
     WeatherView.register_url(app)
     CalendarView.register_url(app)
     MapView.register_url(app)
-
-    # OAuth 2.0 authentication for Google APIs
-    OAuth2View.register_url(app)
-    OAuth2CallbackView.register_url(app)
-    OAuth2RevokeView.register_url(app)
-    Oauth2ClearView.register_url(app)
 
     # Register error handler to known status codes
     error_handler = make_error_handler()
